FILTRO

+++ database/migrations/2025_08_19_024338_add_template_id_to_empresasite_table.php
@@ -20,9 +20,11 @@ return new class extends Migration
 
     public function down(): void
     {
-        Schema::table('empresa_site', function (Blueprint $table) {
-            $table->dropForeign(['template_id']);
-            $table->dropColumn('template_id');
+         Schema::table('empresa_site', function (Blueprint $table) {
+            if (Schema::hasColumn('empresa_site', 'template_id')) {
+                $table->dropForeign(['template_id']);
+                $table->dropColumn('template_id');
+            }
         });
     }
 };
